转发上线下线事件DTO(补充信息：最近30条状态)到kafka

+++ vc-biz/vc-biz-pub-model/src/main/java/com/ccclubs/pub/orm/dto/StateDTO.java
@@ -44,7 +44,7 @@ public class StateDTO implements Serializable {
     /**
      * 小电瓶电量
      */
-    private BigDecimal power;
+    private Integer power;
 
     /**
      * 动力电池电量
@@ -244,11 +244,11 @@ public class StateDTO implements Serializable {
         this.fuelQuantity = fuelQuantity;
     }
 
-    public BigDecimal getPower() {
+    public Integer getPower() {
         return power;
     }
 
-    public void setPower(BigDecimal power) {
+    public void setPower(Integer power) {
         this.power = power;
     }
